Timeline editor: multi-track, audio/video sync, waveform
Added audio waveform rendering, marker-accurate beat/time mapping, and synchronized audio/video playback using NAudio and LibVLCSharp.
Introduced TimelineContextService for tool integration, new converters, and improved serialization.
Includes UI/UX improvements, error handling, and code cleanup. Lays groundwork for advanced timeline editing features.

+++ JustDanceEditor.Formats.Unity/Bundles/MapPackageBundleBuilder.cs
@@ -316,7 +316,7 @@ public static class MapPackageBundleBuilder
         foreach (SectionSegment section in trackStructure.Sections ?? [])
         {
             AssetTypeValueField newSection = ValueBuilder.DefaultValueFieldFromArrayTemplate(sectionsArray);
-            newSection["MusicSection"]["sectionType"].AsInt = section.SectionType;
+            newSection["MusicSection"]["sectionType"].AsInt = (int)section.SectionType;
             newSection["MusicSection"]["marker"].AsDouble = section.StartBeat;
             newSection["MusicSection"]["comment"].AsString = section.Comment ?? string.Empty;
             sectionsArray.Children.Add(newSection);
